requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.1
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
flake8==4.0.1
h11==0.12.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
pylint==2.12.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uvicorn==0.16.0
wrapt==1.13.3
